DDS Group Project/pyproject.toml
[project]
name = "group-distributed-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.7",
    "apache-airflow-providers-google>=20.0.0",
    "bs4>=0.0.2",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pymongo>=4.16.0",
    "pytest>=9.0.2",
    "torch>=2.10.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
